requirements.txt
alembic==1.4.2
appdirs==1.4.3
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cryptography==2.9
Flask==1.1.2
Flask-API==2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
http-ece==1.1.0
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jwt==1.0.0
Mako==1.1.2
MarkupSafe==1.1.1
pep517==0.6.0
Pillow==7.1.1
pkg-resources==0.0.0
py-vapid==1.7.0
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
python-for-android==2020.3.30
pytoml==0.1.21
pywebpush==1.10.1
requests==2.23.0
sh==1.12.14
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.8
virtualenv==16.7.10
Werkzeug==1.0.1
WTForms==2.2.1
zipp==3.1.0
